Moved tests out of source files
