+++ libaurorawallet/src/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod handle_store;
 pub mod multi_pool;
-pub mod test_utils;
