+++ libaurorawallet/tests/test_utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod config;
+pub mod helper_functions;
